Solve and enforce ruff RUF rules
